// src/components/Table/types.ts
export type TypeSearch = {
  date_release: Date | string;
  date_revision?: Date | string;
  description: string;
  id: string;
  logo: string;
  name: string;
};

export type TypeProduct = {
  date_release: Date | string;
  date_revision?: Date | string;
  description: string;
  id: string;
  logo: string;
  name: string;
};